Extraction of project name from REST requests always failed
The code to extract the project name from an /api/projects/*
REST request never seemed to have worked, instead the result was
always "projects" and the lookup in project service consequently
returned always null. Thus, the login with the permit service
was never in the context of a concrete project. That has no
consequences as long as the requesting user is either an admin,
or the REST request is a GET, in which case the project context
is not relevant for a permission check. However, with the new
writable REST API the project context suddenly became relevant
and all permission checks failed.

Change-Id: Ice25a105e2ef3e3e6f48d9ce438cafa5727a272d
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>