Make REST API v2 the default
The XStream-based REST API is going to be retired and the newe API
has been intensively tested for incompatibilities in the meantime.
Time to make v2 the default. The old API is still available by
adding the query parameter rest=v1 to a request.

The test data for RestAPISmokeTest has been adapted accordingly.

Change-Id: Ib7daab098b92f09bc217e512209fa768620e52f5
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com