Support for partial backups
Additional query parameters "from", "to" and "period" are introduced
that allow specifying the time frame for a backup.

"period" gives a time interval in the form <numeral><unit>, e.g.
"5d" or "7h". Supported units are "d" for days, "h" for hours and
"m" for minutes (all case insensitive).

"from" and "to" can be timestamps in ISO format, for instance
"from=2013-10-15T00:00:00Z", or time intervals like "-7d", which
is interpreted as "7 days before now".

"period" can be combined with either "from" or "to" to define, for
example, a certain time interval beginning at a given point in time
in the past. If both "from" and "to" are given, "period" is ignored.

If no time interval for the backup is specified, 1 day is assumed.
For a full backup, now the query parameter "all" must be added.

Change-Id: I70591c11536a7049752430bc8f8f6ff196e86a35
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>