Introduce skalli.destinationService property
In some environments a dedicated destination service must be
used to create connections; in this case the built-in fallback
to DestinationComponent implemented in the Destinations helper
class produced HTTP clients that were unusable for communication.

This patch introduces a new property "skalli.destinationService"
that can be set to specify the destination service that must be
used by the Destinations helper class. Destinations will now
return null if either the specified service is not available,
or it cannot handle the given URL.

HostReachableValidator has been changed to avoid meaningless
issues in case Destinations cannot deliver a suitable client.
Project owners cannot fix these issues anyway.

Change-Id: I1a6cfa14aebaf2d2976197b60b75637957141548
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>